healthai_nexus/requirements.txt
streamlit
langchain
langchain-core
langchain-groq
python-dotenv
pandas
pillow
pytesseract
matplotlib
reportlab
PyPDF2
requests
plotly
